requirements.txt
flask>=3.0.0
gunicorn
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
